src/test/java/com/leo/leetcode/tree/T111Test.java
package com.leo.leetcode.tree;

import com.leo.leetcode.builtin.TreeNode;
import org.junit.Assert;
import org.junit.Test;

import static org.junit.Assert.*;

public class T111Test {

    @Test
    public void minDepth() {
        T111 t111 = new T111();
        TreeNode node = new TreeNode(3);
        node.left = new TreeNode(9);
        node.right = new TreeNode(20);
        node.right.left = new TreeNode(15);
        node.right.right = new TreeNode(7);
        Assert.assertEquals(
                2,
                t111.minDepth(node)
        );
        node = new TreeNode(1);
        node.left = new TreeNode(2);
        node.right = new TreeNode(3);
        node.left.left = new TreeNode(4);
        node.right.right = new TreeNode(5);
        Assert.assertEquals(
                3,
                t111.minDepth(node)
        );
    }
}